Camera functionality is added
I added provider to the manifest (This was needed to be able to read the uri of the captured photo!). I also created a new method which uploads images to the storage (Because the older one actually does not do anything).  Now we are not using the method from the firebaseMethods. We are using the method uploadImageToFirebase which is in the ViewGridItemViewFragment class.